trekk utilities ut av index.ts
