+++ src/utils/assert.ts
@@ -1,0 +1,7 @@
+export const assert = <T extends any>(value: T | undefined | null, msg?: string): T => {
+	if (!value) {
+		throw new Error(msg || 'Value is missing');
+	}
+
+	return value;
+};
