+++ src/utils/auth/auth-token-utils.ts
@@ -1,7 +1,7 @@
 import { TokenSet } from 'openid-client';
-import { assert } from '../index.js';
+import { assert } from '../assert.js';
 import { Request } from 'express';
-import { JsonData } from '../json-utils.js';
+import { JsonData } from '../config-utils.js';
 import { fromBase64 } from '../utils.js';
 import { secureLog } from '../logger.js';
