--- src/utils/json-utils.ts
@@ -1,17 +1,0 @@
-export interface JsonData {
-    [key: string]: any;
-}
-
-const substituteEnvVariables = (_key: string, value: any): any => {
-    if (typeof value !== "string") return value;
-    const matches = value.match('{{(.*?)}}');
-    if (!matches) return value;
-
-    const valueFromEnv = process.env[matches[1]];
-    if (!valueFromEnv) return value;
-    return valueFromEnv;
-}
-
-export function parseJSONwithSubstitutions(text: string): JsonData {
-    return JSON.parse(text, substituteEnvVariables);
-}
